RFC: arm: dom0: allow static memory configuration
 [pure RFC, not for merge]

First of all it was very annoying to check Dom0 address on any cfg change
and update Zephyr dom0 DT - having fixed dom0 RAM and Xen load address
allows to fix "memory" node and "hypervisor" Grant table addresses.

Dom0 is direct mapped by default and, in general, have static memory, but
the memory is allocated from Heap and may have different start addresses.
So, if it exactly known what is required, why not explicitly define where
to load Dom0. It will increase predictability and simplicity of the system.
More over, It will allow to specify and use fixed addresses for DTB and
ramdisk in the future if required (for Zephyr for example).

So, This RFC patch simply allows to specify "xen,static-mem" in "chosen"
node as optinal parameter, and, if specified, allocate requested static
memory for Dom0.

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>